+++ src/solver/gpu_standard_mpm_solver.cu
@@ -192,10 +192,10 @@ namespace crmpm
         if (idx >= numParticles)
             return;
 
-        // Set deformation tensor to identity
-        particleGradientDeformationTensorColumn0[idx].x = 1;
-        particleGradientDeformationTensorColumn1[idx].y = 1;
-        particleGradientDeformationTensorColumn2[idx].z = 1;
+        // Set deformation gradient tensor to identity
+        particleGradientDeformationTensorColumn0[idx] = make_float4(1, 0, 0, 0);
+        particleGradientDeformationTensorColumn1[idx] = make_float4(0, 1, 0, 0);
+        particleGradientDeformationTensorColumn2[idx] = make_float4(0, 0, 1, 0);
 
         standardMpmComputeInitialGridMass<true>(
             gridBoundMin,
@@ -223,10 +223,10 @@ namespace crmpm
         if (idx >= numParticles)
             return;
 
-        // Set deformation tensor to identity
-        particleGradientDeformationTensorColumn0[idx].x = 1;
-        particleGradientDeformationTensorColumn1[idx].y = 1;
-        particleGradientDeformationTensorColumn2[idx].z = 1;
+        // Set deformation gradient tensor to identity
+        particleGradientDeformationTensorColumn0[idx] = make_float4(1, 0, 0, 0);
+        particleGradientDeformationTensorColumn1[idx] = make_float4(0, 1, 0, 0);
+        particleGradientDeformationTensorColumn2[idx] = make_float4(0, 0, 1, 0);
 
         standardMpmComputeInitialVolume(
             gridBoundMin,